Homework 2 Done
Make a method that will for every day of a week print to console your most important activity for that day (and of course, you can just pretend about the activity :-) ). I want you to program this in at least two different ways: using case, and using just if-than-else. You can put days of the week in an array, list, or use individual variables, up to you.